+++ core/internal/ingest/post.js
@@ -4,6 +4,7 @@ const streamSourceFileDao = require('../../stream-source-files/dao')
 const streamSegmentDao = require('../../stream-segments/dao')
 const fileFormatDao = require('../../stream-segments/dao/file-extensions')
 const { sequelize } = require('../../_models')
+const { getSegmentRemotePath } = require('../../stream-segments/bl/segment-file-utils')
 
 const Converter = require('../../../common/converter')
 const ArrayConverter = require('../../../common/converter/array')
@@ -96,8 +97,13 @@ module.exports = function (req, res) {
             .filter((s) => { return !existingSegments.map(e => e.start.toISOString()).includes(s.start.toISOString()) })
             .map((s) => {
               const fileExtensionId = fileExtensionObjects.find(obj => obj.value === s.file_extension).id
+              const path = getSegmentRemotePath({
+                start: s.start,
+                stream_id: streamId
+              })
               return {
                 ...s,
+                path,
                 stream_id: streamId,
                 stream_source_file_id: streamSourceFile.id,
                 file_extension_id: fileExtensionId