+++ core/streams/dao/index.js
@@ -116,7 +116,7 @@ async function query (filters, options = {}) {
   if (filters.names) {
     where.name = {
       [Sequelize.Op.iLike]: {
-        [Sequelize.Op.any]: filters.names.map(n => `${n.replace(/^\*/, '%').replace(/\*$/, '%')}`)
+        [Sequelize.Op.any]: filters.names.map(n => `${n.replace(/(_|%|\\)/g, '\\$1').replace(/^\*/, '%').replace(/\*$/, '%')}`)
       }
     }
   }
@@ -127,7 +127,7 @@ async function query (filters, options = {}) {
       },
       name: {
         [Sequelize.Op.iLike]: {
-          [Sequelize.Op.any]: filters.namesOrIds.map(n => `${n.replace(/^\*/, '%').replace(/\*$/, '%')}`)
+          [Sequelize.Op.any]: filters.namesOrIds.map(n => `${n.replace(/(_|%|\\)/g, '\\$1').replace(/^\*/, '%').replace(/\*$/, '%')}`)
         }
       }
     }